refactor: Remove the last remnants of the V1 login screen
The V1 login screen itself went away in 111d72950, but the scaffolding
that only existed to switch between V1 and V2 stayed behind:

- `LoginView` was a pure passthrough to `LoginViewV2`, plus an unused
  `AppSettings` observation left over from the removed toggle.
- `LoginViewProtocol` existed solely so `LoginView` could pick one of the
  two implementations.
- `LoginViewModel.loginStateValid` had no callers left once V1 was gone.

Fold `LoginViewV2` back into `LoginView` and drop the rest. The two call
sites already spelled it `LoginView`, so they are unchanged.

`LoginView` defaulted `initial` to true while `LoginViewV2` defaulted it
to false, so the merged type keeps `true` to preserve behaviour at the
initial-login call site. The "Initial" preview was passing `initial =
false` and rendering the additional-login variant; it now matches its
name.
