--- swift-paperless/Views/Login/LoginViewV2.swift
@@ -1,212 +1,0 @@
-//
-//  LoginViewV2.swift
-//  swift-paperless
-//
-//  Created by Paul Gessinger on 04.08.2024.
-//
-
-import AppShared
-import Foundation
-import Persistence
-import SwiftUI
-import os
-
-@MainActor
-private struct DetailsView: View {
-  let connectionManager: ConnectionManager
-
-  @Environment(\.dismiss) private var dismiss
-  @Environment(LoginViewModel.self) private var viewModel
-
-  @Environment(IdentityManager.self) private var identityManager
-
-  var body: some View {
-    // Hack
-    @Bindable var viewModel = viewModel
-
-    NavigationStack {
-      Form {
-        SwiftUI.Section {
-          NavigationLink {
-            ExtraHeadersView(headers: $viewModel.extraHeaders)
-          } label: {
-            Label(
-              markdown: .login(.extraHeaders), systemImage: "list.bullet.rectangle.fill"
-            )
-          }
-          NavigationLink {
-            TLSListView(identityManager: identityManager)
-          } label: {
-            Label(localized: .settings(.identities), systemImage: "lock.fill")
-          }
-
-          NavigationLink {
-            LogView(connectionManager: connectionManager)
-          } label: {
-            Label {
-              Text(.settings(.logs))
-                .accentColor(.primary)
-            } icon: {
-              Image(systemName: "text.word.spacing")
-            }
-          }
-        }
-      }
-      .navigationTitle(Text(.login(.detailsTitle)))
-      .navigationBarTitleDisplayMode(.inline)
-
-      .toolbar {
-        ToolbarItem(placement: .navigationBarLeading) {
-          CancelIconButton()
-        }
-      }
-    }
-  }
-}
-
-struct LoginFooterView<Content: View>: View {
-  var systemImage: String
-  var content: () -> Content
-
-  var body: some View {
-    HStack(alignment: .firstTextBaseline) {
-      Image(systemName: systemImage)
-      VStack(alignment: .leading) {
-        content()
-      }
-    }
-    .font(.footnote)
-  }
-}
-
-@MainActor
-struct LoginViewV2: LoginViewProtocol {
-  var connectionManager: ConnectionManager
-  var initial: Bool = false
-
-  @State private var viewModel = LoginViewModel()
-  @State private var identityManager = IdentityManager()
-
-  @Environment(\.dismiss) private var dismiss
-  @EnvironmentObject private var errorController: ErrorController
-
-  @State private var showDetails = false
-  @State private var stage = LoginStage.connection
-  @State private var showSuccessOverlay = false
-
-  private func loginSucceeded(stored: StoredConnection) {
-    Haptics.shared.notification(.success)
-
-    showSuccessOverlay = true
-    Task {
-      try? await Task.sleep(for: .seconds(2))
-      if !initial {
-        dismiss()
-        try? await Task.sleep(for: .seconds(0.2))
-      }
-      do {
-        try connectionManager.login(stored)
-      } catch {
-        // The credentials were valid but the row didn't reach the DB. Say so
-        // rather than dropping the user back to a login screen that looks
-        // like it just ignored them.
-        Logger.shared.error("Storing the logged-in connection failed: \(error)")
-        errorController.push(error: error)
-      }
-    }
-  }
-
-  var body: some View {
-    NavigationStack {
-      VStack {
-        switch stage {
-        case .connection:
-          ConnectionStageView(stage: $stage)
-            .transition(.move(edge: .leading))
-        case .credentials:
-          CredentialsStageView(onSuccess: loginSucceeded)
-            .transition(.move(edge: .trailing))
-        }
-      }
-      .animation(.spring(duration: 0.3), value: stage)
-
-      .navigationBarTitleDisplayMode(.inline)
-
-      .toolbar {
-        if initial {
-          ToolbarItem(placement: .principal) {
-            LogoTitle()
-              .fixedSize()
-          }
-        }
-
-        ToolbarItem(placement: .topBarTrailing) {
-          Button {
-            showDetails = true
-          } label: {
-            Label(
-              String(localized: .login(.moreToolbarButtonLabel)), systemImage: "ellipsis.circle")
-          }
-        }
-      }
-
-      .safeAreaInset(edge: .top) {
-        StageSelection(stage: $stage)
-          .padding(.top, 5)
-          .frame(maxWidth: .infinity)
-          .padding(.bottom)
-      }
-
-      .if(!initial) { view in
-        view
-          .navigationTitle(String(localized: .login(.additionalTitle)))
-
-          .toolbar {
-            ToolbarItem(placement: .topBarLeading) {
-              CancelIconButton()
-            }
-          }
-      }
-    }
-
-    .sheet(isPresented: $showDetails) {
-      DetailsView(connectionManager: connectionManager)
-    }
-
-    .successOverlay(isPresented: $showSuccessOverlay, duration: 2.0) {
-      Text(.login(.success))
-    }
-
-    .environment(viewModel)
-    .environment(identityManager)
-  }
-}
-
-// - MARK: Previews
-
-#Preview("Initial") {
-  LoginViewV2(connectionManager: ConnectionManager(database: try! Database.inMemory()))
-    .environmentObject(ErrorController())
-}
-
-#Preview("Additional") {
-  LoginViewV2(
-    connectionManager: ConnectionManager(database: try! Database.inMemory()), initial: false
-  )
-  .environmentObject(ErrorController())
-}
-
-#Preview("StageSwitch") {
-  @Previewable @State var stage = LoginStage.connection
-
-  return VStack {
-    StageSelection(stage: $stage)
-    Button("Connection") {
-      stage = .connection
-    }
-
-    Button("Credentials") {
-      stage = .credentials
-    }
-  }
-}
